requirement.txt
cvxp==1.1.13
tqdm==4.61.0
numpy==1.20.2
networkx==2.5.1
scikit-learn==0.24.2
torch==1.8.1